# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Callora API",
    "version": "1.0.0",
    "description": "API contract covering Callora backend billing, usage, and developer routes."
  },
  "paths": {
...
    "/api/admin/usage/anomalies": {
      "get": {
        "summary": "List detected usage anomalies",
        "description": "Returns per-API daily usage anomalies for admin review. Usage events are aggregated to per-API daily call counts over the requested window, then each API's days are scored against that API's own baseline using a z-score test; days whose absolute z-score meets the threshold are flagged as spikes or drops. Requires admin authentication and the admin IP allowlist.",
        "security": [
          {
            "adminApiKey": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start of analysis window (ISO-8601). Defaults to 30 days ago.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End of analysis window (ISO-8601). Defaults to now.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "description": "Absolute z-score at/above which a day is flagged.",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 10,
              "default": 3
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of anomalies to return (most severe first).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "apiId",
            "in": "query",
            "description": "Restrict analysis to a single API ID.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detected anomalies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "anomalies": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "apiId": {
                                "type": "string"
                              },
                              "day": {
                                "type": "string",
                                "example": "2026-03-21"
                              },
                              "type": {
                                "type": "string",
                                "enum": [
                                  "spike",
                                  "drop"
                                ]
                              },
                              "calls": {
                                "type": "integer"
                              },
                              "revenue": {
                                "type": "string"
                              },
                              "baselineMean": {
                                "type": "number"
                              },
                              "stdDev": {
                                "type": "number"
                              },
                              "zScore": {
                                "type": "number"
                              }
                            }
                          }
                        },
                        "summary": {
                          "type": "object",
                          "properties": {
                            "window": {
                              "type": "object",
                              "properties": {
                                "from": {
                                  "type": "string",
                                  "format": "date-time"
                                },
                                "to": {
                                  "type": "string",
                                  "format": "date-time"
                                }
                              }
                            },
                            "threshold": {
                              "type": "number"
                            },
                            "minDataPoints": {
                              "type": "integer"
                            },
                            "seriesAnalyzed": {
                              "type": "integer"
                            },
                            "anomalyCount": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized admin request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden by admin IP allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/usage/{developerId}": {
      "get": {
        "summary": "Inspect a developer usage aggregate",
        "description": "Returns a redacted aggregate snapshot for a developer's live usageStore state. Requires admin authentication and the admin IP allowlist.",
        "security": [
          {
            "adminApiKey": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "developerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage aggregate snapshot retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsageSnapshotResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized admin request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden by admin IP allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usage aggregate not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-api-key"
      }
    },
    "schemas": {
...
      "AdminUsageSnapshotResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminUsageSnapshot"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "PAYMENT_REQUIRED",
          "TOO_MANY_REQUESTS",
          "CONFLICT",
          "INTERNAL_SERVER_ERROR",
          "BAD_GATEWAY",
          "SERVICE_UNAVAILABLE",
          "GATEWAY_TIMEOUT",
          "VALIDATION_ERROR",
          "INVALID_BODY",
          "INVALID_QUERY",
          "INVALID_PARAMS",
          "INVALID_VALUE",
          "GATEWAY_AUTH_CONTEXT_MISSING",
          "UPSTREAM_TARGET_BLOCKED",
          "INSUFFICIENT_BALANCE",
          "SOROBAN_RPC_TIMEOUT",
          "SOROBAN_RPC_ERROR",
          "BILLING_DEDUCTION_FAILED",
          "BILLING_REQUEST_NOT_FOUND",
          "DEVELOPER_NOT_FOUND",
          "API_ACCESS_FORBIDDEN",
          "API_KEY_NOT_FOUND",
          "API_KEY_FORBIDDEN",
          "MISSING_REFRESH_TOKEN",
          "INVALID_REFRESH_TOKEN",
          "REVOKED_TOKEN",
          "EXPIRED_TOKEN",
          "REFRESH_FAILED",
          "REVOKE_FAILED",
          "NOT_AUTHENTICATED",
          "TOKEN_INFO_FAILED",
          "VAULT_NOT_FOUND",
          "VAULT_BALANCE_RETRIEVAL_FAILED",
          "MISSING_AMOUNT",
          "INVALID_AMOUNT_TYPE",
          "INVALID_AMOUNT_FORMAT",
          "INVALID_NETWORK",
          "NETWORK_MISMATCH",
          "INVALID_SOURCE_ACCOUNT",
          "INVALID_TRANSACTION_INPUT",
          "SOURCE_ACCOUNT_NOT_FOUND",
          "INVALID_CONTRACT_ID",
          "NETWORK_UNAVAILABLE",
          "TRANSACTION_BUILD_FAILED",
          "INTERNAL_ERROR",
          "INVALID_WEBHOOK_REGISTRATION",
          "INVALID_WEBHOOK_EVENT_TYPES",
          "WEBHOOK_NOT_FOUND",
          "INVALID_WEBHOOK_URL",
          "WEBHOOK_URL_VALIDATION_FAILED",
          "MISSING_WEBHOOK_SIGNATURE_HEADERS",
          "INVALID_WEBHOOK_TIMESTAMP",
          "WEBHOOK_TIMESTAMP_OUT_OF_WINDOW",
          "MALFORMED_WEBHOOK_SIGNATURE",
          "INVALID_WEBHOOK_SIGNATURE",
          "INVALID_DELIVERY_ID",
          "INVALID_RETRY_POLICY",
          "DLQ_ENTRY_NOT_FOUND",
          "INVALID_IP_FORMAT",
          "IP_NOT_ALLOWED",
          "DATABASE_NOT_AVAILABLE",
          "IDEMPOTENCY_CONFLICT",
          "IDEMPOTENCY_IN_PROGRESS",
          "SIMULATION_FAILED",
          "INVALID_AUTH_HEADER",
          "MISSING_TOKEN",
          "INVALID_TOKEN",
          "MISSING_CLAIMS",
          "TOKEN_EXPIRED",
          "TOKEN_NOT_ACTIVE",
          "QUOTA_REQUEST_NOT_FOUND",
          "QUOTA_REQUEST_ALREADY_RESOLVED",
          "INVALID_QUOTA_REQUEST",
          "REQUEST_TIMEOUT",
          "REQUEST_BODY_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "UNPROCESSABLE_ENTITY",
          "USAGE_AGGREGATE_NOT_FOUND",
          "INVALID_EXPORT_SCHEDULE",
          "EXPORT_SCHEDULE_NOT_FOUND",
          "MISSING_AUTH_FIELDS",
          "AUTH_NOT_IMPLEMENTED",
          "COMPONENT_NOT_CONFIGURED"
        ],
        "description": "Canonical Callora backend error code."
      },
...
    }
  }
}